requirements.txt
psycopg2-binary>=2.9
matplotlib>=3.8
numpy>=1.26
